user me response

+++ src/modules/users/application/use-cases/delete-profile-image.use-case.ts
@@ -21,23 +21,30 @@ export class DeleteProfileImageUseCase {
     private readonly getUserProfile: GetUserProfileUseCase,
   ) {}
 
-  async execute(userId: number, lang: SupportedLang): Promise<void> {
-    const user = await this.getUserProfile.execute(userId, lang);
+ async execute(userId: number, lang: SupportedLang): Promise<void> {
+  const user = await this.userRepo.findById(userId); 
 
-    if (!user.hasProfileImage() || !user.profileImgPublicId) {
-      throw new BadRequestException({
-        success: false,
-        message: t('noimagetodelete', lang),
-      });
-    }
-
-    try {
-      await this.cloudinary.deleteFile(user.profileImgPublicId);
-    } catch (err) {
-      console.warn(' Failed deleting Cloudinary image', err);
-    }
-
-    user.removeProfileImage();
-    await this.userRepo.deleteProfileImage(userId);
+  if (!user) {
+    throw new BadRequestException({
+      success: false,
+      message: t('userNotFound', lang),
+    });
   }
+
+  if (!user.hasProfileImage() || !user.profileImgPublicId) {
+    throw new BadRequestException({
+      success: false,
+      message: t('noimagetodelete', lang),
+    });
+  }
+
+  try {
+    await this.cloudinary.deleteFile(user.profileImgPublicId);
+  } catch (err) {
+    console.warn('Failed deleting Cloudinary image', err);
+  }
+
+  user.removeProfileImage();
+  await this.userRepo.deleteProfileImage(userId);
+}
 }
